+++ SpringEEIT69-5Mart/src/main/java/com/ispan/eeit69/controller/RegController.java
@@ -57,25 +57,6 @@ public class RegController {
 			model.addAttribute("login", "fail");
 			return "/homePage";
 		} else {
-//			Integer memeberId = result.getId();
-//			System.out.println("會員的id" + memeberId);
-//			TeacherPicture result2 = teacherPictureService.findById(2);
-//			TeacherPicture result2 = teacherPictureService.findByMember(result);
-//			if(result2!=null) {
-//			Blob pic =  result2.getPhoto();
-////			System.out.println(pic);
-//			// 將Blob數據轉換為Base64編碼的字符串
-//			byte[] imageBytes = null;
-//			try {
-//				imageBytes = pic.getBytes(1, (int) pic.length());
-//			} catch (SQLException e) {
-//				e.printStackTrace();
-//			}
-//			String base64Image = Base64.getEncoder().encodeToString(imageBytes);
-//			model.addAttribute("base64Image", base64Image);
-//			session.setAttribute("base64Image", base64Image);
-//			}
-			
 			TeacherPicture result2 = teacherPictureService.findByMember(result);
 			System.out.println("搜尋會員的照片=" + result2);
 			if (result2 == null) {
